% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeo7}
\alias{yeo7}
\title{Yeo 7-Network Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
yeo7
}
\description{
Yeo 7-Network Cortical Atlas
}
\keyword{datasets}
