% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegYeo2011-package.R
\docType{package}
\name{ggsegYeo2011-package}
\alias{ggsegYeo2011}
\alias{ggsegYeo2011-package}
\title{ggsegYeo2011: Yeo 2011 Atlas for the 'ggsegverse' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Yeo 2011 resting-state network atlases for the 'ggseg' ecosystem. Provides 'ggseg_atlas' objects with 2D polygon geometry for use with 'ggseg'. Includes both 7-network and 17-network parcellations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegYeo2011}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegYeo2011/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
