% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeo17}
\alias{yeo17}
\title{Yeo 17-Network Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
yeo17
}
\description{
Yeo 17-Network Cortical Atlas
}
\keyword{datasets}
