% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{yeo17}
\alias{yeo17}
\title{Yeo 17-Network Cortical Atlas}
\usage{
yeo17()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Yeo 17-Network Cortical Atlas
}
\seealso{
Other ggseg_atlases: 
\code{\link{yeo7}()}
}
\concept{ggseg_atlases}
