% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{yeo7}
\alias{yeo7}
\title{Yeo 7-Network Cortical Atlas}
\usage{
yeo7()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Yeo 7-Network Cortical Atlas
}
\seealso{
Other ggseg_atlases: 
\code{\link{yeo17}()}
}
\concept{ggseg_atlases}
