% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{yeo7}
\alias{yeo7}
\title{Yeo 7-Network Cortical Atlas}
\usage{
yeo7()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Yeo et al. (2011) 7-network resting-state cortical
parcellation. Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
yeo7()
plot(yeo7())
}
\references{
Yeo BTT, Krienen FM, Sepulcre J, et al. (2011).
The organization of the human cerebral cortex estimated by intrinsic
functional connectivity. \emph{Journal of Neurophysiology}, 106(3):1125-1165.
\doi{10.1152/jn.00338.2011}
}
\seealso{
Other ggseg_atlases: 
\code{\link{yeo17}()}

Other cortical_atlases: 
\code{\link{yeo17}()}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
